Changes book titles (#3438)
**Description**

Changes "de Hobbit" to "The Hobbit" for all languages except NL, and updates the capitalization of "Green Eggs and Ham".

**Fixes #3437 **

**How to test**

Verify that the book titles have changed in the quiz of level 3, question 7.

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [x] Describes changes in the format above (present tense)
- [x] Links to an existing issue or discussion 
- [x] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!