[TESTING] Attempt 2 for fixing test optimization (#3461)
First, provide a one line summary of your changes in the **Title of the PR**

**Now fill out the remainder of this template by replacing all _italic_ content**

**Description**

_Changes in detail_

For example: "**Adds** translations of levels 1 to 12 to Polish" (note the use of present tense!)

**Fixes _issue or discussion number_**

Always link the number of the issue or of the discussion that your pr concerns.

**How to test**

* If this is a UI change: _describe how to run your code to see it in action. See this https://github.com/Felienne/hedy/pull/880#issue-1016304308 for an example_
* If this is a language change: _add a few tests showing the difference_

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Describes changes in the format above (present tense)
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!