[Language] Allows floats and expressions with turtle (#3481)
**Description**

Changes the grammar and the transpiler to allow floats and expressions for turtle commands. Also, changes a changes a few tests due to them working just for ints. So they are limited now until level 11.

**Fixes #3345*

**How to test**
I added some tests to test this behaviour, with latin and non-latin numbers

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Describes changes in the format above (present tense)
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!