[BUG FIX] Incorrect inputs correctly highlighted again multiple accounts creation (#3494)
We make a bit of an ugly work-around that requires some explanation. For styling of input attributes we have some global classes set. In this case the input:required and input:required:valid classes are relevant. When an input is required but not valid (yet), the input:required styling is set: setting a red border on any input element. When the input is :valid we overwrite this styling with a green border using !important.

This is what is causing the issue. On the moment of receiving the faulty feedback from our POST request we style each input element matching the input with the class border-red-500. But as the element is already styling with a green border through the !important suffix the overwrite doesn't work! In this PR we make a bit of dirty workaround where we remove the required attribute from the inputs at the moment of making the POST request. This doesn't have any functional implications as all verification is handled by the server. However, the border will no longer turn green after the first attempt.

Co-authored-by: Felienne <felienne@gmail.com>