[BUG FIX] Admin is able to set user tags again (#3495)
**Description**
Fix an issue where Dynamo would be mad when sending a dictionary that also contained the index key to the database for altering an entry. Locally this was not an issue so this was hard to detect.

**Fixes**
No related issue.

**How to test**
Deploy and verify that you, as an admin, can set user tags again.