[BUG] Fixes suffixes when displaying variable names (#3513)
**Description**

When using a variable name that's in Javascript reserved words defined by Skuplt, it adds a suffix $rw$. In this PR we'll substract that suffix to display for the user.

**Fixes #3432**

**How to test**
You can go to any level starting from level 2 and use a variable name like name, switch, do, in, etc.

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Describes changes in the format above (present tense)
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!