Extend list of reserved words (#3462)
**Description**

Adds 'input' to the list of reserved words, as it's a Python method that our generated code can call.

**Fixes issues/3454**

Fixes https://github.com/Felienne/hedy/issues/3454 (I think and hope)

**How to test**

This program fails on level 11 before the change but should succeed after it:
```
input = "some string"
x = ask "question"
```

If we want, I can try to add a unit-test for this, but I'm not sure where to put it. `test_level_02.py` since the the issue can be reproduced from that level onward? (Or perhaps it's a bit overkill to test every reserved keyword? I'm flexible on the matter.)

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [x] Describes changes in the format above (present tense)
- [x] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!