[IMPROVEMENT] Changes highest level reached from highest saved program to highest level quiz score (#3503)
**Description**
As pointed out by @JimJamShazam our current implementation of "Highest level reached" is a bit outdated and still depends on relic code from when we didn't do any user logging (and didn't have any more stats than "saved programs"). In PR we improve this by iterating the user logs and getting the highest level for which a quiz is finished.

Note: We don't look into the actual quiz score, only that it is finished.

**Fixes**
This PR fixes #3501.

**How to test**
Make sure you have a class with students in it. Verify that the highest level reached column now depends on the highest level for which a quiz is finished instead of the highest level a program is saved in.