[LOCALIZATION] Implements localized 'search' string for ace search box (#3576)
**Description**
In this PR we make a bit of an ugly work-around to localize the 'search' string in the ace search box. As we can only inject Jinja variables inside our HTML templates we create an empty div that contains our string. Then we replace the placeholder of our search box with the innerHTML of this div: making it localized. However, there is also another solution and I don't really have a preference. Two options:

- Create hidden div with localized string, replace placeholder in `ext-searchbox.js`
- Create minor script in `incl-editor-and-output.html` and overwrite placeholder

Note that in this PR we fix this issue with the first workaround. However, this does leave a hidden div in our template.

**Fixes**
This PR fixes #3536.