[CHORE] Add database with data for local testing (#3571)
**Description**

Adds a database with data to feed the local database when needed. This file does not replace the local database, instead I added a shell script that copies the content from one to the other, avoiding merge conflicts and so. Also updates the CONTRIBUTING.md file to reflect the contents of this database.

**Fixes #3448**