Fixed the variable Prix in FR quiz question (#3589)
There is a variable name "prix" is the same for plural and singular in the French translation. The correction was the same as the original and we couldn't do the quiz question.

In this PR it is changed to "coût" which has a different plural like "price" in English. 

Also, in the same question there was a repeated price that is fixed.