[CONTENT] #3259 reorder adventures based on difficulty (#3628)
**Now fill out the remainder of this template by replacing all _italic_ content**

**Description**
Work towards #3259 
All adventures in level 1 - 13 are now reordered so they go from easiest to hardest. I also added a lot of blanks in the example codes and added challenges to each adventure, so the students know what to do with it and not just play around with the example. 

I still need to do this for level 14 -18, but that's a job for next week