[TESTING] Only test snippets when PR comes from Weblate (#3656)
**Description**
In this PR we expand our Github actions file to only run the snippet tests when a PR comes from Weblate. As we know this PR can only contain content changes we should be fine by only validating the code snippets. This highly speeds up the process of testing. We do have to keep in mind that, if for some reason, the Weblate translations break the front-end we don't test this any longer.

**Fixes**
This PR fixes #3652.

**How to test**
Merge this PR, re-run the jobs for a weblate PR and verify that only the snippets are tested.