[FIX] Fixes issue with french front-end error when running program (#3674)
**Description**
In this PR we fix an issue with the `runit()` function that was broken in French. This was due to the default error message `already_programming_running` containing an `'` character. As the placeholder was also surrounded by `' '` to clarify as a string for Typescript the function broke. We fix this by using the Jinja2 function `|tojson` to always safely escape html special characters.

**Fixes**
This PR fixes #3661

**How to test**
Verify that Hedy works again in French.