[FIX] Fix highlighting to use unicode categories (#3690)
**Description**

Changes the regex used to detect a character and the end of a word in the highliter, preventing problems with it not detecting end of word correctly in languages that don't use the Latin alphabet.

**Fixes #3476**

**How to test**

Change language to hebrew and try this code in level 5:
```
לכל מספר בתוך טווח 1 עד 5
    הדפס מספר
```

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Contains one of the PR categories in the name
- [ ] Describes changes in the format above
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!