[TESTING, CYPRESS] 3692 front-end test for the run code button and programmers mode switch (#3702)
[CYPRESS, TESTING]

**Description**
With these front-end tests the 'Run code' and programmers mode switch on the `/hedy` page are tested. I created id's in the HTML file to select the buttons and I implemented a function which navigates to the `/hedy` page that can be used for these tests but also for future tests.

**Fixes _issue or discussion number_**
Fixes #3692 

**How to test**
Run
`export ADMIN_USER=admin_user; python app.py` in your virtual environment. This ensures that there is an active admin_user and Hedy is running.

Then you can go to the /tests folder and run:

`npx cypress run --spec cypress/e2e/hedy_page/`