[CLEANUP] Make Hedy PEP8 compliant (#3602)
**Description**

Since #3410 we introduced a git pre-commit hook to lint modified code. However, much of the Hedy codebase is still not PEP8 compliant, in this PR we'll change that by applying autopep8 to the whole codebase with the following parameters:

```
autopep8 --in-place --max-line-length 120  -aaa --jobs 0 [files]
```
A few changes have been made to the pre-commit hook:
* The max line length now is 120
* I excluded the max line check from utils.py  because the number dictionary is too long.

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Describes changes in the format above (present tense)
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!