[TESTING, CYPRESS] 3704 testing creating front end test for coding on the hedy page (#3754)
[TESTING]

**Description**
These front-end tests tests many features on the `/hedy` page. I was unable to test the read aloud button because it did not show in Cypress and thus I was unable to select it. This pull requests tests the editor box, the 'try' button, back to level button, next level button and switching between adventures in a level.

**Fixes _issue or discussion number_**
Fixes #3704 

**How to test**
These tests uses the accounts in the test database, to use this the database has to be fed first by running bash `feed_dev_database.sh` in the hedy (root) directory.
To run the tests on the command, do the the following: `npx cypress run --spec tests/cypress/e2e/hedy_page`
This tests all the hedy page tests.