[UI] Add overlay to disabled level button  (#3711)
**Description**
Fixes issue in #3431. Adds an overlay on hover to the disabled level button to give more information on why the level button is disabled. 



**How to test**

* If this is a UI change: _describe how to run your code to see it in action. See this https://github.com/Felienne/hedy/pull/880#issue-1016304308 for an example_

If you log in as a teacher and then go to customizations and disable some of the levels (either by setting the quiz threshold or locking a level), you can then log in as a student and see that the "Go to level _" button is disabled (as well as the previous level button) with the correct tool tip explaining why it is disabled. 

<img width="1078" alt="Screenshot 2022-11-30 at 10 30 38 PM" src="https://user-images.githubusercontent.com/59891012/204959547-5657bc5f-b3cf-4ce4-b738-9fa17a883ce3.png">
<img width="1077" alt="Screenshot 2022-11-30 at 10 30 03 PM" src="https://user-images.githubusercontent.com/59891012/204959552-6598bc39-04cc-48e8-9dc9-fa93945197b8.png">
<img width="1076" alt="Screenshot 2022-11-30 at 10 29 50 PM" src="https://user-images.githubusercontent.com/59891012/204959554-6c65be41-b058-41f4-8f95-565edaee81de.png">

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Contains one of the PR categories in the name
- [ ] Describes changes in the format above
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!