[TESTING, CYPRESS] 3765 added an id for the adventure classes and use this to select the buttons (#3781)
[TESTING]

**Description**
There is now id for every adventure tab in the `/hedy` which is used to select the tabs. To check if the tab is selected I use the class of the element.

**Fixes _issue or discussion number_**
Fixes #3765

**How to test**
These tests uses the accounts in the test database, to use this the database has to be fed first by running bash `feed_dev_database.sh` in the hedy (root) directory.
To run the tests on the command, do the the following: `npx cypress run --spec tests/cypress/e2e/hedy_page`
This tests all the hedy page tests.