[FIX] Fixes error where /explore filtering would always fail (#3790)
**Description**
In this PR we fix quite some bugs that occurred on the `/explore` page when filtering. An incorrect return was set when the `default` adventure was chosen for filtering, resulting in an incorrect return response. We also fix an issue where for the `code_lines` on the `Hedy choice` programs a variable `code` was used that didn't exist. Lastly we improve the parsing of the keywords dynamically and the error checking that wasn't performed on the Hedy choice programs.

This is changed into a function and called in both loops to make sure everything works as expected. To make sure this keeps working it would be nice to add some e2e tests on the filtering. If I recall correctly we did have some of these, but apparently the above described bugs weren's caught.

**Fixes**
This PR fixes #3785 and fixes #3786.

**How to test**
Verify that you can filter again on the `/explore` page. Make sure the filtering for `default` adventure works as expected.