[FIX] i escaping and modified tests accordingly - bug #3676 (#3804)
** Fixes #3676 **

I changed the variable "i" to be __i __. This change causes i to not show up on the variable drop down. I also had to change test suite 7 by replacing all of the "i's" with __i __ so that they would work correctly. 


Note: I'm not 100% sure why changing i to __i __ fixes this bug, I was messing around with changing the variable and it just worked. So, this could be an incorrect approach. 

**How to test**
On level 7: Story, you can see that without the fix, the variable "i" shows up in the variable drop down. With the fixed code, the "i" variable doesn't show up. 

Link: https://hedy.org/hedy/7#story

<img width="1004" alt="Screen Shot 2022-12-07 at 6 55 03 PM" src="https://user-images.githubusercontent.com/59888688/206322632-4a722323-c4a7-4117-a60c-2c58e7384520.png">
<img width="997" alt="Screen Shot 2022-12-07 at 6 54 24 PM" src="https://user-images.githubusercontent.com/59888688/206322651-6616882a-d15b-43f7-a78e-ebdbce7274e3.png">