3778 testing two buttons on classpage (#3843)
**Description**

This pull request adds test for the customize_class button and the go_back_to_teacher_page button. On the  `/for-teacher/class/[class id]` page.

**Fixes _issue or discussion number_**
Fixes #3778

**How to test**
These tests uses the accounts in the test database, to use this the database has to be fed first by running bash `feed_dev_database.sh` in the `hedy` (root) directory.
To run the tests on the command, do the the following: `npx cypress run --spec tests/cypress/e2e/for-teacher_page/class_page`
This tests all the hedy page tests.