[CHORE] Clean-up styles.css by removing hard-coded CSS classes (#3840)
**Description**
We clean-up the tooltip overlay implemented on the next level button when this is disabled. It is currently written with native CSS and some not really nice styling enforcements. We attempt to clean this up by depending solely on Tailwind and removing redundant styling attributes. It is still not ideal, but works good enough. To improve the UX we should rely on Javascript for the dynamic interaction.

**Fixes**
This PR fixes #3783

**How to test**
Verify that the tooltip overlay is still working (however, a bit re-styled).