[LANGUAGE] Adds buttons to Hedy (#3748)
**Description**

Adds a new rule to the grammar to make it possible to define a button, with the syntax:
```python
hello is button
```

Changes the grammar rule ```ifpressed``` so that instead of a keyboard character the name of a button can be used, with the expected outcome.

Changes the front-end HTML templates to add a div containing the user created buttons just below the output screen.

**Fixes**

This PR fixes #3678.

Always link the number of the issue or of the discussion that your PR concerns.
Tip, if you use the words fixes before the issue number in this description, the related issue will automatically close then the PR is merged! 

**How to test**

In level 5 run the following Hedy code:
```python
hello is button

if hello is pressed print "Hello world!"
```
A button should appear, and pressing the button should result in the string `Hello world!` appearing in the output and the program terminating.

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [x] Contains one of the PR categories in the name
- [x] Describes changes in the format above
- [x] Links to an existing issue or discussion 
- [x] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!