[FIX] Fixing unit test (level 7) (#3864)
Menno noted that the last unit test of level 7 did not have an assert, which let to a test that will not be used. 

To fix this, I added the missing assertion in unit test 'test_if_pressed_repeat' and made sure the unit test does not fail. This PR therefor fixes issue #3863.