[FIX] Refactor /expore page
Refactor the database access of the /explore page. We now
use key-only indices to filter the database records (potentially
still not using the most effective database access mechanism,
but should be better), and use batch get to retrieve the public
profile information. Also put a timeout on it.

All this effort for nothing -- turns out the time is mostly
spent in translating the keywords in the examples. This invokes
at least one and potentially 2 Hedy parses for every program on
the explore page.

Still committing this because it should be better, but the real
work is elsewhere.