[FIX] Prevent server overload by achievement pushing
During log research, we've observed the `/push-achivement` endpoint
being hit ~2000 times in 2 minutes. This may be a programming bug
that we're not quite sure how it's triggered yet.

Protect the backend by not pushing the same achivement more than once
in a single page load; so that even if this is runaway JavaScript
that does `pushAchievement()` in a loop, at least we don't hit
the backend after the first time.