[FIX] Remove startup translation of keywords in YAMLs
The eager keyword translation upon startup adds about ~1 minute
to the startup time of the server, and complicates the code.

Replace with on-demand translation. After timing, this seems to add
10-15ms to the very worst page load: loading all adventures. That seems
like an acceptable price to pay to simplify the code a lot, improve
startup time, and improve memory usage as well since we don't have to
keep multiple copies of the YAML data in memory.

Reduction of memory usage may lead to us being able to run more
processes in parallel on the server, which will be good for
responsiveness.