[REFACTOR] Reduce memory usage by not querying Babel
The Python library `babel` wraps `gettext` and a locale database.

However, when the information about a locale is accessed, they choose to
cache that information in memory indefinitely. We load information about
all locales at startup, which fills the cache with data we'll never read
again.

This accounts for at least 40MB of memory usage of the server process (probably more).

Instead, we will read the bits of information we need from the Babel locale
database in a build tool, and write it to a file; we will then read that
file in the server process.